Move FontGPUBuffer class to new Velaptor.OpenGL.Buffers namespace & folder
