+++ Velaptor/OpenGL/Buffers/FontGPUBuffer.cs
@@ -2,7 +2,7 @@
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.OpenGL
+namespace Velaptor.OpenGL.Buffers
 {
     // ReSharper disable RedundantNameQualifier
     using System;
